feat(engine): extract content-lane's pure leaf modules to loopover-engine
Move the two dependency-free leaf modules of the content-lane subtree —
content-repo-spec.ts (the curated-list registry spec) and flag.ts (the
content-lane enablement flag) — into packages/loopover-engine, leaving thin
re-export shims at their old src/ paths so every consumer and test resolves
unchanged. Follows the established engine-extraction convention exactly (see
src/signals/check-summary.ts): the shim re-exports the engine implementation via
a relative source path, not the published package.

These two files are pure (no imports, no Env/D1 dependency) and already at 100%
coverage, so the move is coverage-safe. This validates the shim pattern on the
cleanest boundary before extracting the interdependent and Env-bound files.

Closes #4880
